moved override keybindings to job status widget; using override shortcut now shows override panel
