+++ src/app/widgets/JobStatus/JobStatus.jsx
@@ -25,14 +25,20 @@
 import PropTypes from 'prop-types';
 import React, { PureComponent } from 'react';
 import get from 'lodash/get';
+import pubsub from 'pubsub-js';
 import reduxStore from 'app/store/redux';
+import combokeys from 'app/lib/combokeys';
+import controller from 'app/lib/controller';
+import log from 'app/lib/log';
 import { UPDATE_JOB_OVERRIDES } from 'app/actions/visualizerActions';
 import { connect } from 'react-redux';
 import TooltipCustom from 'app/components/TooltipCustom/ToolTip';
 import ToggleSwitch from 'app/components/ToggleSwitch';
+import { OVERRIDES_CATEGORY, OVERRIDE_VALUE_RANGES } from '../../constants';
 import IdleInfo from './components/IdleInfo';
 import Overrides from './components/Overrides';
 import styles from './index.styl';
+import useKeybinding from '../../lib/useKeybinding';
 
 /**
  * Job Status component wrapper
@@ -42,6 +48,200 @@ class JobStatus extends PureComponent {
     static propTypes = {
         state: PropTypes.object,
     };
+
+    shuttleControlFunctions = {
+        FEEDRATE_OVERRIDE: (_, { amount }) => {
+            const { fileLoaded, connection, ovF } = this.props;
+            const feedRate = ovF + (Number(amount) || 0);
+
+            if (fileLoaded && connection.isConnected) {
+                if (get(reduxStore.getState(), 'visualizer.jobOverrides.toggleStatus') === 'jobStatus') {
+                    this.handleOverrideToggle();
+                }
+                if (feedRate <= OVERRIDE_VALUE_RANGES.MAX && feedRate >= OVERRIDE_VALUE_RANGES.MIN) {
+                    switch (Number(amount)) {
+                    case 1:
+                        controller.write('\x93');
+                        break;
+                    case -1:
+                        controller.write('\x94');
+                        break;
+                    case 10:
+                        controller.write('\x91');
+                        break;
+                    case -10:
+                        controller.write('\x92');
+                        break;
+                    case 0:
+                        controller.write('\x90');
+                        break;
+                    default:
+                        break;
+                    }
+                    pubsub.publish('feedrate:change', feedRate);
+                } else {
+                    log.error('ovF out of range: ' + this.props.ovF);
+                }
+            }
+        },
+        SPINDLE_OVERRIDE: (_, { amount }) => {
+            const { fileLoaded, connection, ovS } = this.props;
+            const spindleSpeed = ovS + (Number(amount) || 0);
+
+            if (fileLoaded && connection.isConnected) {
+                if (get(reduxStore.getState(), 'visualizer.jobOverrides.toggleStatus') === 'jobStatus') {
+                    this.handleOverrideToggle();
+                }
+                if (spindleSpeed <= OVERRIDE_VALUE_RANGES.MAX && spindleSpeed >= OVERRIDE_VALUE_RANGES.MIN) {
+                    switch (Number(amount)) {
+                    case 1:
+                        controller.write('\x9C');
+                        break;
+                    case -1:
+                        controller.write('\x9D');
+                        break;
+                    case 10:
+                        controller.write('\x9A');
+                        break;
+                    case -10:
+                        controller.write('\x9B');
+                        break;
+                    case 0:
+                        controller.write('\x99');
+                        break;
+                    default:
+                        break;
+                    }
+                    pubsub.publish('spindlespeed:change', spindleSpeed);
+                } else {
+                    log.error('ovS out of range: ' + this.props.ovS);
+                }
+            }
+        },
+    }
+
+    shuttleControlEvents = {
+        FEEDRATE_OVERRIDE_P: {
+            title: 'Feed +',
+            keys: '',
+            gamepadKeys: '5',
+            keysName: 'R1',
+            cmd: 'FEEDRATE_OVERRIDE_P',
+            payload: { amount: 1 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
+        },
+        FEEDRATE_OVERRIDE_PP: {
+            title: 'Feed ++',
+            keys: '',
+            gamepadKeys: '',
+            keysName: 'Feed ++',
+            cmd: 'FEEDRATE_OVERRIDE_PP',
+            payload: { amount: 10 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
+        },
+        FEEDRATE_OVERRIDE_M: {
+            title: 'Feed -',
+            keys: '',
+            gamepadKeys: '7',
+            keysName: 'R2',
+            cmd: 'FEEDRATE_OVERRIDE_M',
+            payload: { amount: -1 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
+        },
+        FEEDRATE_OVERRIDE_MM: {
+            title: 'Feed --',
+            keys: '',
+            gamepadKeys: '',
+            keysName: 'Feed --',
+            cmd: 'FEEDRATE_OVERRIDE_MM',
+            payload: { amount: -10 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
+        },
+        FEEDRATE_OVERRIDE_RESET: {
+            title: 'Feed Reset',
+            keys: '',
+            gamepadKeys: '',
+            keysName: 'Feed Reset',
+            cmd: 'FEEDRATE_OVERRIDE_RESET',
+            payload: { amount: 0 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
+        },
+        SPINDLE_OVERRIDE_P: {
+            title: 'Spindle/Laser +',
+            keys: '',
+            gamepadKeys: '4',
+            keysName: 'L1',
+            cmd: 'SPINDLE_OVERRIDE_P',
+            payload: { amount: 1 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
+        },
+        SPINDLE_OVERRIDE_PP: {
+            title: 'Spindle/Laser ++',
+            keys: '',
+            gamepadKeys: '',
+            keysName: 'Spindle/Laser ++',
+            cmd: 'SPINDLE_OVERRIDE_PP',
+            payload: { amount: 10 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
+        },
+        SPINDLE_OVERRIDE_M: {
+            title: 'Spindle/Laser -',
+            keys: '',
+            gamepadKeys: '6',
+            keysName: 'L2',
+            cmd: 'SPINDLE_OVERRIDE_M',
+            payload: { amount: -1 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
+        },
+        SPINDLE_OVERRIDE_MM: {
+            title: 'Spindle/Laser --',
+            keys: '',
+            gamepadKeys: '',
+            keysName: 'Spindle/Laser --',
+            cmd: 'SPINDLE_OVERRIDE_MM',
+            payload: { amount: -10 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
+        },
+        SPINDLE_OVERRIDE_RESET: {
+            title: 'Spindle/Laser Reset',
+            keys: '',
+            gamepadKeys: '',
+            keysName: 'Spindle/Laser Reset',
+            cmd: 'SPINDLE_OVERRIDE_RESET',
+            payload: { amount: 0 },
+            preventDefault: true,
+            isActive: true,
+            category: OVERRIDES_CATEGORY,
+            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
+        },
+    }
 
     /**
      * Determine the file size format between bytes, kilobytes (KB) and megabytes (MB)
@@ -74,6 +274,22 @@ class JobStatus extends PureComponent {
         reduxStore.dispatch({ type: UPDATE_JOB_OVERRIDES, payload: JSON.parse(localStorage.getItem('jobOverrideToggle')) });
     }
 
+    addShuttleControlEvents() {
+        combokeys.reload();
+
+        Object.keys(this.shuttleControlEvents).forEach(eventName => {
+            const callback = eventName === 'MACRO' ? this.shuttleControlEvents[eventName] : this.shuttleControlEvents[eventName].callback;
+            combokeys.on(eventName, callback);
+        });
+    }
+
+    removeShuttleControlEvents() {
+        Object.keys(this.shuttleControlEvents).forEach(eventName => {
+            const callback = eventName === 'MACRO' ? this.shuttleControlEvents[eventName] : this.shuttleControlEvents[eventName].callback;
+            combokeys.removeListener(eventName, callback);
+        });
+    }
+
     componentDidUpdate() {
         if (!this.props.fileLoaded || !this.props.connection.isConnected) {
             localStorage.setItem('jobOverrideToggle', JSON.stringify({ isChecked: false,
@@ -86,6 +302,12 @@ class JobStatus extends PureComponent {
             isChecked: false,
             toggleStatus: 'jobStatus',
         }));
+        this.addShuttleControlEvents();
+        useKeybinding(this.shuttleControlEvents);
+    }
+
+    componentWillUnmount() {
+        this.removeShuttleControlEvents();
     }
 
     render() {
@@ -144,10 +366,17 @@ export default connect((store) => {
     const filteredPath = path.replace(name, '');
     const connection = get(store, 'connection');
     const activeState = get(store, 'controller.state.status.activeState', 'Idle');
+    const overrides = get(store, 'controller.state.status.ov', [0, 0, 0]);
+
+    const ovF = overrides[0];
+    const ovS = overrides[2];
+
     return {
         ...file,
         filteredPath,
         connection,
-        activeState
+        activeState,
+        ovF,
+        ovS,
     };
 })(JobStatus);
