+++ src/app/widgets/Visualizer/index.jsx
@@ -72,9 +72,7 @@ import {
     GRBL_ACTIVE_STATE_CHECK,
     CARVING_CATEGORY,
     GENERAL_CATEGORY,
-    OVERRIDES_CATEGORY,
     VISUALIZER_CATEGORY,
-    OVERRIDE_VALUE_RANGES
 } from '../../constants';
 import {
     CAMERA_MODE_PAN,
@@ -839,60 +837,6 @@ class VisualizerWidget extends PureComponent {
     }
 
     shuttleControlFunctions = {
-        FEEDRATE_OVERRIDE: (_, { amount }) => {
-            const feedRate = this.props.ovF + (Number(amount) || 0);
-            if (feedRate <= OVERRIDE_VALUE_RANGES.MAX && feedRate >= OVERRIDE_VALUE_RANGES.MIN) {
-                switch (Number(amount)) {
-                case 1:
-                    controller.write('\x93');
-                    break;
-                case -1:
-                    controller.write('\x94');
-                    break;
-                case 10:
-                    controller.write('\x91');
-                    break;
-                case -10:
-                    controller.write('\x92');
-                    break;
-                case 0:
-                    controller.write('\x90');
-                    break;
-                default:
-                    break;
-                }
-                pubsub.publish('feedrate:change', feedRate);
-            } else {
-                log.error('ovF out of range: ' + this.props.ovF);
-            }
-        },
-        SPINDLE_OVERRIDE: (_, { amount }) => {
-            const spindleSpeed = this.props.ovS + (Number(amount) || 0);
-            if (spindleSpeed <= OVERRIDE_VALUE_RANGES.MAX && spindleSpeed >= OVERRIDE_VALUE_RANGES.MIN) {
-                switch (Number(amount)) {
-                case 1:
-                    controller.write('\x9C');
-                    break;
-                case -1:
-                    controller.write('\x9D');
-                    break;
-                case 10:
-                    controller.write('\x9A');
-                    break;
-                case -10:
-                    controller.write('\x9B');
-                    break;
-                case 0:
-                    controller.write('\x99');
-                    break;
-                default:
-                    break;
-                }
-                pubsub.publish('spindlespeed:change', spindleSpeed);
-            } else {
-                log.error('ovS out of range: ' + this.props.ovS);
-            }
-        },
         VISUALIZER_VIEW: (_, { type }) => {
             const {
                 to3DView,
@@ -1066,126 +1010,6 @@ class VisualizerWidget extends PureComponent {
                     this.workflowControl.handleOnStop();
                 }
             },
-        },
-        FEEDRATE_OVERRIDE_P: {
-            title: 'Feed +',
-            keys: '',
-            gamepadKeys: '5',
-            keysName: 'R1',
-            cmd: 'FEEDRATE_OVERRIDE_P',
-            payload: { amount: 1 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
-        },
-        FEEDRATE_OVERRIDE_PP: {
-            title: 'Feed ++',
-            keys: '',
-            gamepadKeys: '',
-            keysName: 'Feed ++',
-            cmd: 'FEEDRATE_OVERRIDE_PP',
-            payload: { amount: 10 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
-        },
-        FEEDRATE_OVERRIDE_M: {
-            title: 'Feed -',
-            keys: '',
-            gamepadKeys: '7',
-            keysName: 'R2',
-            cmd: 'FEEDRATE_OVERRIDE_M',
-            payload: { amount: -1 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
-        },
-        FEEDRATE_OVERRIDE_MM: {
-            title: 'Feed --',
-            keys: '',
-            gamepadKeys: '',
-            keysName: 'Feed --',
-            cmd: 'FEEDRATE_OVERRIDE_MM',
-            payload: { amount: -10 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
-        },
-        FEEDRATE_OVERRIDE_RESET: {
-            title: 'Feed Reset',
-            keys: '',
-            gamepadKeys: '',
-            keysName: 'Feed Reset',
-            cmd: 'FEEDRATE_OVERRIDE_RESET',
-            payload: { amount: 0 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.FEEDRATE_OVERRIDE,
-        },
-        SPINDLE_OVERRIDE_P: {
-            title: 'Spindle/Laser +',
-            keys: '',
-            gamepadKeys: '4',
-            keysName: 'L1',
-            cmd: 'SPINDLE_OVERRIDE_P',
-            payload: { amount: 1 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
-        },
-        SPINDLE_OVERRIDE_PP: {
-            title: 'Spindle/Laser ++',
-            keys: '',
-            gamepadKeys: '',
-            keysName: 'Spindle/Laser ++',
-            cmd: 'SPINDLE_OVERRIDE_PP',
-            payload: { amount: 10 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
-        },
-        SPINDLE_OVERRIDE_M: {
-            title: 'Spindle/Laser -',
-            keys: '',
-            gamepadKeys: '6',
-            keysName: 'L2',
-            cmd: 'SPINDLE_OVERRIDE_M',
-            payload: { amount: -1 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
-        },
-        SPINDLE_OVERRIDE_MM: {
-            title: 'Spindle/Laser --',
-            keys: '',
-            gamepadKeys: '',
-            keysName: 'Spindle/Laser --',
-            cmd: 'SPINDLE_OVERRIDE_MM',
-            payload: { amount: -10 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
-        },
-        SPINDLE_OVERRIDE_RESET: {
-            title: 'Spindle/Laser Reset',
-            keys: '',
-            gamepadKeys: '',
-            keysName: 'Spindle/Laser Reset',
-            cmd: 'SPINDLE_OVERRIDE_RESET',
-            payload: { amount: 0 },
-            preventDefault: true,
-            isActive: true,
-            category: OVERRIDES_CATEGORY,
-            callback: this.shuttleControlFunctions.SPINDLE_OVERRIDE
         },
         VISUALIZER_VIEW_3D: {
             title: '3D / Isometric',
@@ -1604,16 +1428,12 @@ export default connect((store) => {
     const controllerType = get(store, 'controller.type');
     const activeState = get(store, 'controller.state.status.activeState');
     const alarmCode = get(store, 'controller.state.status.alarmCode');
-    const overrides = get(store, 'controller.state.status.ov', [0, 0, 0]);
     const isFileLoaded = get(store, 'file.fileLoaded');
 
     const { activeVisualizer } = store.visualizer;
 
     const feedArray = [xMaxFeed, yMaxFeed, zMaxFeed];
     const accelArray = [xMaxAccel * 3600, yMaxAccel * 3600, zMaxAccel * 3600];
-
-    const ovF = overrides[0];
-    const ovS = overrides[2];
 
     return {
         feedArray,
@@ -1625,8 +1445,6 @@ export default connect((store) => {
         activeState,
         activeVisualizer,
         alarmCode,
-        ovF,
-        ovS,
         isFileLoaded
     };
 }, null, null, { forwardRef: true })(VisualizerWidget);
