+++ src/package.json
@@ -1,6 +1,6 @@
 {
   "name": "gSender",
-  "version": "1.2.1",
+  "version": "1.2.2",
   "productName": "gSender",
   "description": "CNC Milling Controller",
   "author": {
